+++ mybatis-page/src/main/java/software/sitb/spring/data/mybatis/dialect/H2Dialect.java
@@ -6,7 +6,7 @@ public class H2Dialect extends Dialect {
     private final static String LIMIT_SQL_PATTERN_FIRST = "%s limit %s";
 
     @Override
-    public String getLimitString(String sql, int offset, int limit) {
+    public String getLimitString(String sql, long offset, long limit) {
         if (offset == 0) {
             return String.format(LIMIT_SQL_PATTERN_FIRST, sql, limit);
         }
